Revert "Make those NSIntegers so some warnings go away"
This reverts commit 5ae34fc3df32e87dd62bd564423d2f9d4c3a4d92.

+++ Quicksilver/Code-External/CGSPrivate.h
@@ -53,9 +53,9 @@
 */
 
 // Internal CoreGraphics typedefs
-typedef NSInteger	CGSConnection;
-typedef NSInteger   CGSWindow;
-typedef NSInteger   CGSValue;
+typedef int		CGSConnection;
+typedef int		CGSWindow;
+typedef int		CGSValue;
 
 //// CONSTANTS ////
